Renamed components folder to features and added javascript unit tests
