+++ src/app/app-routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
-import { GuestBookComponent } from './components/guest-page/guest-book/guest-book.component';
-import { PostsPageComponent } from './components/blog-posts/posts-page/posts-page.component';
-import { PostDetailsPageComponent } from './components/blog-posts/post-details-page/post-details-page.component';
+import { GuestBookComponent } from './features/guest-page/guest-book/guest-book.component';
+import { PostsPageComponent } from './features/blog-posts/posts-page/posts-page.component';
+import { PostDetailsPageComponent } from './features/blog-posts/post-details-page/post-details-page.component';
 
 export const appRoutes: Routes = [
   { path: '', redirectTo: '/blog-posts', pathMatch: 'full' },
