+++ src/app/features/guest-page/author-popup/author-popup.component.ts
