+++ src/app/features/guest-page/models/guest-entry.model.ts
