+++ src/app/features/guest-page/services/avatar.service.ts
